requirements.txt
Flask==3.0.2
flask-restx==1.3.0
RPi.GPIO==0.7.1
smbus==1.1.post2
spidev==3.6
